Remove VMs that should not be deleted
This takes output from a powercli script that finds VMs in a "to be deleted" list that are either powered on or don't exist on the server powercli is connected to. I needed a way to remove the "bad" vms from the master list so I could proceed with deleting the correct ones.